Updated configs and other stuff
Rewrote the whole "adm_config.php" to be more accurate in describing all
features and functions.

Implemented function for seperate IP for the server and MySQL, also
added option for MySQL port.

Implemented the custom title defined as "SITENAME".

Updated the footer information to include UltraTM's name and added link
to github page.